Wizard: Rename `component` directory to `components`
This renames `component` directory to `components` for the sake of consistency.
